Putting the boat destroyed Modal, and threating some nulls

+++ BattleShip/src/it/unical/asde/battleship/components/controllers/LoginController.java
@@ -48,8 +48,9 @@ public class LoginController
     @GetMapping("/logout")
     public String logout(final HttpSession session)
     {
-        utilService.deletePlayingUser((User) session.getAttribute("user"));
-
+    	if(session.getAttribute("user")!=null) {
+    		utilService.deletePlayingUser((User) session.getAttribute("user"));
+    	}
         session.invalidate();
         return "redirect:/";
     }
